requirements.txt
Flask==0.10.1
github3.py==1.0.0a4
celery==3.1.23
gunicorn==0.17.2
argparse>=1.2.0,<=1.3.0
cryptography==2.1.4
PyJWT==1.5.3
# linters
jinja2<2.9
pep8>=1.7.0,<2.0.0
flake8>=3.3.0,<3.4.0
yamllint>=1.6.1,<1.7.0
demjson>=2.2.3,<2.3.0
ansible-lint>=3.4.11,<3.5.0
pylint>=1.7.4
